Functional update
1. The overflow of unsigned integers in the vector storing the
pair-counts is now avoided by using long long unsinged int.

2. Name of the executable file now changed into "kdtpcf".

+++ driver.h
@@ -48,9 +48,9 @@ private:
 private:                        // Data
     kdtree data,      rand;
     int    data_size, rand_size;
-    std::vector<unsigned> dd, dd_jk;
-    std::vector<unsigned> rr, rr_jk;
-    std::vector<unsigned> dr, dr_jk;
+    std::vector<long long unsigned> dd, dd_jk;
+    std::vector<long long unsigned> rr, rr_jk;
+    std::vector<long long unsigned> dr, dr_jk;
 private:                        // Function
     void cal(  );
     void cal_ls(  );			// Landy-Sazlay
